my_python/requirements.txt
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0rc1
charset-normalizer==3.3.2
ciscoconfparse==1.9.51
configparser==7.0.0
cryptography==43.0.0
DateTime==5.5
Deprecated==1.2.14
dnspython==2.6.1
future==1.0.0
hier-config==2.2.3
idna==3.7
ipaddress==1.0.23
loguru==0.7.2
netmiko==4.4.0
ntc_templates==6.0.0
packaging==23.2
paramiko==3.4.0
passlib==1.7.4
pycparser==2.22
PyNaCl==1.5.0
pynetbox==7.3.4
pyserial==3.5
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scp==0.15.0
six==1.16.0
textfsm==1.1.3
toml==0.10.2
urllib3==2.2.2
wrapt==1.16.0
zope.interface==6.4.post2
